requirements.txt
fastapi~=0.100.1
openai~=0.27.8
panel~=1.2.1
python-dotenv~=1.0.0
langchain~=0.0.251